requirements.txt
linkpreview==0.1.11
requests==2.22.0
tinydb==4.1.1
Flask==1.1.1
filetype==1.0.7
python-magic==0.4.18
shortuuid==1.0.1
cryptography~=3.0